requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
